Created Authenticateable as an interface to the Authenticate package. Create and Login class implements all four methods from the interface, but only use two each. The other two are implemented to return false. Might not be the best solution.

+++ ElectroshoppenProjekt/src/Authentication/Create.java
@@ -36,10 +36,14 @@ public class Create extends Authentication {
     }
 
     /**
-     *
-     * @return
+     * Method checks whether or not the user exists in the database, and saves it
+     * there if not.
+     * @param type Must be customer or employee, else an IllegalArgumentException is thrown
+     * @return True if user succesfullt created and saved to the database.
+     * false if user already exists in the database.
+     * 
      */
-    public String createUser(String type) {
+    public boolean createUser(String type) {
 
 	//Create new Profile instance based on given type
 	String typeLower = type.toLowerCase();
@@ -52,38 +56,76 @@ public class Create extends Authentication {
 		    this.eMail, this.adress, super.password);
 
 	} else {
-	    return "No such type of profile.";
-	}
-	//Ask the instance to save itself to the database    
-
-	if (!super.userExists()) {
-	    this.profile.saveProfileToDB();
-	    return "Should be saved";
-
-	} else {
-	    return "Profile already exists";
-	}
-    }
-
-    public String deleteUser(String type) {
-	String typeLower = type.toLowerCase();
-	String query = "";
-	if (typeLower.equals("customer")){
-	    query = "DELETE FROM customer WHERE password = '" 
-		    + super.password + "' AND phone_number = '" 
-		    + super.userName + "';";
-	} else if (typeLower.equals("employee")){
-	    query = "DELETE FROM customer WHERE password = '" 
-		    + super.password + "' AND phone_number = '" 
-		    + super.userName + "';";
-	} else {
-	    return "No such type of profile";
+	    throw new IllegalArgumentException("Type must be customer or employee");
 	}
 	
+        //Ask the instance to save itself to the database    
+	if (!super.userExists()) {
+	    this.profile.saveProfileToDB();
+	    return true;
+
+	} else {
+	    return false;
+	}
+    }
+    
+    
+    /**
+     * 
+     * @param type Must be customer or employee, else an IllegalArgumentException is thrown
+     * @return True if the user is succesfully deleted from the database.
+     * false if the user does not exist in the database.
+     */
+    public boolean deleteUser(String type) {
+	String typeLower = type.toLowerCase();
+	String query = "";
+        
+        if (super.userExists()) {
+            
+            //Delete customer profile query
+            if (typeLower.equals("customer")){
+                query = "DELETE FROM customer WHERE password = '" 
+                        + super.password + "' AND phone_number = '" 
+                        + super.userName + "';";
+                
+            //Delete employee profile query  
+            } else if (typeLower.equals("employee")){
+                query = "DELETE FROM customer WHERE password = '" 
+                        + super.password + "' AND phone_number = '" 
+                        + super.userName + "';";
+            } else {
+                throw new IllegalArgumentException("Type must be customer or employee");
+            }
+        } else {
+            return false;
+        }
+	
+        //Actually run the query
 	DBConnection dbc = new DBConnection();
 	dbc.runQueryUpdate(query);
-	return "Profile deleted succesfully";
+	return true;
 
+    }
+    
+    
+    /**
+     * Do not use this method.
+     * Meant for Login class.
+     * @return false
+     */
+    @Override
+    public boolean doLogin() {
+        return false;
+    }
+    
+    /**
+     * Do not use this method.
+     * Meant for Login class.
+     * @return false
+     */
+    @Override
+    public boolean doLogout() {
+        return false;
     }
 
 }
